Update massage.py
Added tools for building json blobs for creating new entries (from a ledger_line, the tabular structure that ledgerize kicks out). For now, we only support JournalEntries.

I also added a names_list maker, which is basically a chart of accounts but for Vendor, Customer, and Employee objects. (I basically looked for the intersection of all those objects' attributes.)